Fixations

+++ sip-creator/src/main/java/eu/delving/sip/files/StorageImpl.java
@@ -598,7 +598,9 @@ public class StorageImpl extends StorageBase implements Storage {
                         if (progressListener != null) progressListener.finished(!cancelled);
                         outputStream.close();
                         File hashedSource = new File(here, hasher.prefixFileName(FileType.SOURCE.getName()));
-                        FileUtils.moveFile(source, hashedSource);
+                        if (!hashedSource.exists()) {
+                            FileUtils.moveFile(source, hashedSource);
+                        }
                     }
                     else {
                         File file = new File(here, fileName);
